refactor(coordinator): extract APNs code-attest from api/provider.go
Move the v0.6.0 APNs code-identity attestation flow (codeAttestMetric,
codeAttestLoop, tryCrossVersionReuse, maybeRearmCodeAttest,
sendCodeIdentityChallenge, handleCodeAttestationResponse) into
api/provider_codeattest.go. Same-package move, no behavior change;
provider.go 3563 -> 3150 LOC. Full api go test ./api/... stays green.
